chore(parquet): reader refactoring
